+++ server/impl/src/main/java/com/walmartlabs/concord/server/org/project/ProjectDao.java
@@ -135,7 +135,7 @@ public class ProjectDao extends AbstractDao {
                 return null;
             }
 
-            Result<Record11<UUID, UUID, String, String, String, String, String, Boolean, UUID, String, String>> repos = tx.select(
+            Result<Record10<UUID, UUID, String, String, String, String, String, UUID, String, String>> repos = tx.select(
                     REPOSITORIES.REPO_ID,
                     REPOSITORIES.PROJECT_ID,
                     REPOSITORIES.REPO_NAME,
@@ -143,7 +143,6 @@ public class ProjectDao extends AbstractDao {
                     REPOSITORIES.REPO_BRANCH,
                     REPOSITORIES.REPO_COMMIT_ID,
                     REPOSITORIES.REPO_PATH,
-                    REPOSITORIES.HAS_WEBHOOK,
                     SECRETS.SECRET_ID,
                     SECRETS.SECRET_NAME,
                     SECRETS.STORE_TYPE)
@@ -153,7 +152,7 @@ public class ProjectDao extends AbstractDao {
                     .fetch();
 
             Map<String, RepositoryEntry> m = new HashMap<>();
-            for (Record11<UUID, UUID, String, String, String, String, String, Boolean, UUID, String, String> repo : repos) {
+            for (Record10<UUID, UUID, String, String, String, String, String, UUID, String, String> repo : repos) {
                 m.put(repo.get(REPOSITORIES.REPO_NAME),
                         new RepositoryEntry(
                                 repo.get(REPOSITORIES.REPO_ID),
@@ -165,7 +164,6 @@ public class ProjectDao extends AbstractDao {
                                 repo.get(REPOSITORIES.REPO_PATH),
                                 repo.get(SECRETS.SECRET_ID),
                                 repo.get(SECRETS.SECRET_NAME),
-                                repo.get(REPOSITORIES.HAS_WEBHOOK),
                                 repo.get(SECRETS.STORE_TYPE)));
             }