+++ server/impl/src/main/java/com/walmartlabs/concord/server/repository/RepositoryProvider.java
@@ -29,5 +29,5 @@ public interface RepositoryProvider {
 
     void fetch(UUID orgId, RepositoryEntry repository, Path dest);
 
-    RepositoryManager.RepositoryInfo getInfo(Path path);
+    RepositoryInfo getInfo(Path path);
 }