+++ server/impl/src/test/java/com/walmartlabs/concord/server/template/ProjectDaoTest.java
@@ -67,7 +67,7 @@ public class ProjectDaoTest extends AbstractDaoTest {
 
         String repoName = "repo#" + System.currentTimeMillis();
         String repoUrl = "n/a";
-        repositoryDao.insert(projectId, repoName, repoUrl, null, null, null, null, false);
+        repositoryDao.insert(projectId, repoName, repoUrl, null, null, null, null);
 
         // ---
         Map<String, Object> newCfg1 = ImmutableMap.of("a1", "a1-v");